feat(chat-deploy): added chat to deploy flow
