+++ sim/app/w/[id]/layout.tsx
@@ -1,13 +1,9 @@
-import { Chat } from './components/chat/chat'
 import { ErrorBoundary } from './components/error'
 
 export default function WorkflowLayout({ children }: { children: React.ReactNode }) {
   return (
-    <>
-      {/* <Chat /> */}
-      <main className="bg-muted/40 overflow-hidden h-full">
-        <ErrorBoundary>{children}</ErrorBoundary>
-      </main>
-    </>
+    <main className="bg-muted/40 overflow-hidden h-full">
+      <ErrorBoundary>{children}</ErrorBoundary>
+    </main>
   )
 }
